Used random number generator to generate sleep
Used random number generator to generate sleep. 

To test if the threading is being done appropriately, I used random number generator to generate random number for sleep. 
It actually capture the completed and the still running sleep process while thread is still being limited to thread_num. 

However, the script will exit if loop has reached its last data and one of the process in pids completed.  

NEXT:
Find a way on not to exit while there are still pid in pids.